+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
 ## 0.0.1 (October 2, 2018)
 - init project
-- add essential part (init config & create from template)
+- add essential part (init config & create from template & run certain code)
+- only C and CPP is supported